+++ source_gen/source_gen/lib/builder.dart
@@ -2,4 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+/// Import 'source_gen.dart' instead, which exports `GeneratorBuilder` as well.
+///
+/// Will be removed in _0.6.0_.
+@deprecated
+library source_gen.builder;
+
 export 'src/builder.dart';